Refactor FileManager with SOLID principles
